Pientä hiontaa

+++ src/main/java/ohtu/radioaine/database/DAO.java
@@ -2,7 +2,7 @@ package ohtu.radioaine.database;
 
 public interface DAO<T> {
     public T create(T instance);
-    public T read(int id);
+    public T read(long id);
     public void delete(T instance);
     public T update(T instance);
 }
